+++ CHANGELOG.md
@@ -20,6 +20,7 @@ In addition to original Keep-a-Changelog, we use following rules:
 ### Fixed
 - `SUBTYPE_CONSTRAINT` cannot parse supertype-constraint like `ONEOF` https://github.com/ricosjp/ruststep/pull/205
 ### Internal
+- Use Instantiables instead of Constraint https://github.com/ricosjp/ruststep/pull/206
 - Return index from ir::Namespace::resolve https://github.com/ricosjp/ruststep/pull/204
 - cargo clippy --fix https://github.com/ricosjp/ruststep/pull/202
 - Index named AST https://github.com/ricosjp/ruststep/pull/201